fix: schema examples should account for $ref objects
Currently the recursive function that generates the example
codeblocks for reference data types does not account for a
function call where there is no definition.type value.

In the case of a $ref to another JSONschema, the object will
look like:

{
  "$ref" : "path-to-other-schema.json"
}

Though the datatables functions accounted for $ref, the
examples did not. This commit corrects this oversight.