refactor: schema reference data to include patternProperties (#97)
In JSONSchema "patternProperties" properties represent property
names that match a provided regex pattern. In Hamlet schemas this
is used to reference part of a JSON structure that is to be
defined by the file's author. Most commonly this is the object
"id".

Such properties are very important to understanding the expected
JSON structure, but have thus far not been accounted for in
examples on the website. This PR introduces an asterisk "*" to
represent patternProperties, indicating to the user that this is
necessary but self-defined.

Co-authored-by: Ross Murray <rossmurray@protonmail.com>