docs: hamlet learn - targetted hamlet tutorials
Hamlet Learn is to house all of the "tutorial"-type documentation
for the Hamlet Deploy project. These guided practical lessons
will provide exposure to the fundamental concepts of Hamlet,
whilst leaving the conceptual/reference/discussion documentation
for other sections of the site.

The tutorials are deliberately written to be approachable, fun and
not too serious, whilst not getting bogged down in the explanations.