chore: remove non-default github templates
The hamlet-io github org has a .github repository that provides
default repository documentation such as for issue and PRs.

They only apply to a given repository within the org when there
is none already in the repository.

This commit removes these non-default ones from the repo so that
the defaults can apply and can be consistent across the org.