refactor: split large recursive datatables function into two (#114)
The getDataTables function is unwieldy and hard to maintain. It's recursive,
however the data structures are different at several levels. The function
returns an array of datatables, with pre-processed attributes for rows.

This refactor replaces the single function with two:

* getDataTablesList
* getDataTableRows

getDataTablesList determines the dataTables that will be required for a schema,
and for each table it returns the table name, and its raw data.

getDataTableRows is then used to process this raw data and determine the table
rows required (schema attributes) - it does not need to recurse into children
because that will be in another dataTable.

Together, these functions are far simpler to maintain.

Co-authored-by: Ross Murray <rossmurray@protonmail.com>